Add files via upload
This is already a completed project for the Scholarship Chair. This takes in a Questionnaire put out by the VP of Comms, organizes the classes and names associated with the classes and outputs a clean file with a representation of key, value pairs in a table kind of like Class List.xlsx.